+++ internal/client/src/client/client_test.go
@@ -19,8 +19,8 @@ import (
 	"time"
 
 	"github.com/SIGBlockchain/project_aurum/internal/constants"
-	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accounts"
 
+	"github.com/SIGBlockchain/project_aurum/internal/producer/src/accounts/accountstable"
 	"github.com/SIGBlockchain/project_aurum/internal/producer/src/block"
 	producer "github.com/SIGBlockchain/project_aurum/internal/producer/src/producer"
 	"github.com/SIGBlockchain/project_aurum/pkg/keys"
@@ -576,7 +576,7 @@ func TestRequestWalletInfo(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			if tt.name == "Wallet address in table" {
-				err = accounts.InsertAccountIntoAccountBalanceTable(dbc, walletAddress, tt.expectedBal)
+				err = accountstable.InsertAccountIntoAccountBalanceTable(dbc, walletAddress, tt.expectedBal)
 				if err != nil {
 					t.Errorf("failed to insert account into account balance table")
 				}
